# openapi.yaml
openapi: 3.0.3
info:
  title: 医养结合云端后台管理系统 API
  version: 2.0.0
  description: |
    基于 train.md 与当前 Spring Boot 实现维护的 REST API 契约。
    除登录、公开注册和刷新令牌外，接口均需要 Bearer access token。
    JSON 成功响应的业务 code 固定为 0，错误响应同时使用正确的 HTTP 状态码。
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: 登录
      operationId: login
      security: []
      requestBody:
        $ref: '#/components/requestBodies/Login'
      responses:
        '200': { $ref: '#/components/responses/LoginResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/register:
    post:
      tags: [Auth]
      summary: 公开注册老人账号
      description: 公开注册固定创建 elderly 角色账号，并返回 access token 与可轮换 refresh token。
      operationId: register
      security: []
      requestBody:
        $ref: '#/components/requestBodies/Register'
      responses:
        '201': { $ref: '#/components/responses/LoginResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: 刷新并轮换令牌
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshTokenRequest' }
      responses:
        '200': { $ref: '#/components/responses/LoginResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      tags: [Auth]
      summary: 获取当前用户
      operationId: getCurrentUser
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: 退出登录并使当前 access token 失效
      operationId: logout
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LogoutRequest' }
      responses:
        '200': { $ref: '#/components/responses/EmptyResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /assessment-reports:
    get:
      tags: [AssessmentReport]
      summary: 评估报告列表
      operationId: listAssessmentReports
      parameters:
        - { name: elderlyId, in: query, schema: { type: string } }
        - { $ref: '#/components/parameters/PageNo' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiAssessmentReportPage' }
    post:
      tags: [AssessmentReport]
      summary: 生成评估报告
      description: 未提供评分、等级、摘要、风险项或建议时，后端根据有效预警计算生成。
      operationId: createAssessmentReport
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssessmentReportCreateRequest' }
      responses:
        '201': { $ref: '#/components/responses/AssessmentReportResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /assessment-reports/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [AssessmentReport]
      summary: 报告详情
      operationId: getAssessmentReport
      responses:
        '200': { $ref: '#/components/responses/AssessmentReportResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [AssessmentReport]
      summary: 软删除报告
      operationId: deleteAssessmentReport
      responses:
        '200': { $ref: '#/components/responses/EmptyResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /assessment-reports/{id}/review:
    post:
      tags: [AssessmentReport]
      summary: 复核评估报告
      operationId: reviewAssessmentReport
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssessmentReportReviewRequest' }
      responses:
        '200': { $ref: '#/components/responses/AssessmentReportResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /assessment-reports/{id}/export:
    get:
      tags: [AssessmentReport]
      summary: 导出 Markdown 报告文件
      operationId: exportAssessmentReport
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: UTF-8 Markdown 文件
          headers:
            Content-Disposition:
              schema: { type: string }
          content:
            text/markdown:
              schema: { type: string, format: binary }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /devices:
    get:
      tags: [Device]
      summary: 设备列表
      operationId: listDevices
      parameters:
        - { name: bindingStatus, in: query, schema: { type: string, enum: [bound, unbound] } }
        - { name: onlineStatus, in: query, schema: { type: string, enum: [online, offline] } }
        - { $ref: '#/components/parameters/PageNo' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiDevicePage' }
    post:
      tags: [Device]
      summary: 新增设备
      operationId: createDevice
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceCreateRequest' }
      responses:
        '201': { $ref: '#/components/responses/DeviceResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /devices/bind:
    post:
      tags: [Device]
      summary: 绑定设备到老人档案
      operationId: bindDevice
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceBindRequest' }
      responses:
        '200': { $ref: '#/components/responses/DeviceResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /devices/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Device]
      summary: 设备详情
      operationId: getDevice
      responses:
        '200': { $ref: '#/components/responses/DeviceResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Device]
      summary: 更新设备
      operationId: updateDevice
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceUpdateRequest' }
      responses:
        '200': { $ref: '#/components/responses/DeviceResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Device]
      summary: 软删除设备
      operationId: deleteDevice
      responses:
        '200': { $ref: '#/components/responses/EmptyResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /devices/{id}/unbind:
    post:
      tags: [Device]
      summary: 解绑设备
      operationId: unbindDevice
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/DeviceResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /devices/{id}/reports:
    get:
      tags: [Device]
      summary: 查询最近 100 条设备上报
      operationId: getDeviceReports
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/DeviceReportListResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Device]
      summary: 写入设备数据上报
      operationId: recordDeviceReport
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceDataReportRequest' }
      responses:
        '201': { $ref: '#/components/responses/DeviceReportResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /ai/analysis/care-decision:
    post:
      tags: [AIAnalysis]
      summary: 生成并持久化照护决策分析
      description: 当前实现使用可复现的业务规则引擎，根据请求中的结构化指标生成结果并持久化。
      operationId: createCareDecisionAnalysis
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnalysisRequest' }
      responses:
        '200': { $ref: '#/components/responses/AnalysisResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /ai/analysis/care-decision/{id}:
    get:
      tags: [AIAnalysis]
      summary: 分析详情
      operationId: getCareDecisionAnalysis
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/AnalysisResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, maxLength: 64 }
...
    PageNo:
      name: pageNo
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
  requestBodies:
    Login:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/LoginRequest' }
    Register:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/RegisterRequest' }
...
  responses:
    EmptyResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiEmpty' }
    LoginResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiLogin' }
    UserResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiUser' }
...
    AssessmentReportResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiAssessmentReport' }
...
    DeviceResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiDevice' }
...
    DeviceReportResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiDeviceReport' }
    DeviceReportListResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiDeviceReportList' }
...
    AnalysisResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiAnalysis' }
...
    BadRequest:
      description: 参数或业务输入错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: 未登录、令牌无效或账号已禁用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: 当前角色无权访问
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: 数据不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: 数据冲突或仍被引用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    ApiBase:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: success }
        traceId: { type: string, example: b3c1f8e9d8f14a7f }
    ApiEmpty:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { nullable: true }
    ApiError:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: integer, enum: [400, 401, 403, 404, 409, 500, 501] }
        message: { type: string }
        data: { nullable: true }
        traceId: { type: string }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, writeOnly: true }
    RegisterRequest:
      type: object
      required: [username, password, realName]
      properties:
        username: { type: string, minLength: 3, maxLength: 50, pattern: '^[A-Za-z0-9_]+$' }
        password: { type: string, minLength: 8, maxLength: 64, writeOnly: true }
        realName: { type: string, maxLength: 50 }
        mobile: { type: string, pattern: '^1[3-9]\d{9}$' }
        roleCode: { type: string, enum: [elderly], default: elderly }
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string, writeOnly: true }
    LogoutRequest:
      type: object
      properties:
        refreshToken: { type: string, writeOnly: true }
    LoginResult:
      type: object
      required: [token, accessToken, refreshToken, tokenType, expireAt, accessExpireAt, refreshExpireAt, user]
      properties:
        token: { type: string, description: 兼容旧前端的 accessToken 别名 }
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, example: Bearer }
        expireAt: { type: string, format: date-time, description: 兼容旧前端的 accessExpireAt 别名 }
        accessExpireAt: { type: string, format: date-time }
        refreshExpireAt: { type: string, format: date-time }
        user:
          $ref: '#/components/schemas/LoginUserInfo'
...
    ApiLogin:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/LoginResult' }
    RoleCode:
      type: string
      enum: [admin, doctor, operator, analyst, elderly]
    User:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        realName: { type: string }
        roleCode: { $ref: '#/components/schemas/RoleCode' }
        mobile: { type: string, nullable: true }
        institutionId: { type: string, nullable: true }
        regionCode: { type: string, nullable: true }
        status: { type: string, enum: [enabled, disabled] }
        lastLoginAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ApiUser:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/User' }
...
    AssessmentReport:
      type: object
      properties:
        id: { type: string }
        elderlyId: { type: string }
        reportType: { type: string }
        score: { type: integer, minimum: 0, maximum: 100 }
        grade: { type: string }
        summary: { type: string }
        riskItems: { type: array, items: { type: string } }
        recommendations: { type: array, items: { type: string } }
        assessorId: { type: string, nullable: true }
        assessedAt: { type: string, format: date-time }
        reviewStatus: { type: string, enum: [draft, approved, rejected] }
        reviewerId: { type: string, nullable: true }
        reviewedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AssessmentReportCreateRequest:
      type: object
      required: [elderlyId, reportType]
      properties:
        elderlyId: { type: string, maxLength: 64 }
        reportType: { type: string, maxLength: 50 }
        score: { type: integer, minimum: 0, maximum: 100 }
        grade: { type: string, maxLength: 10 }
        summary: { type: string, maxLength: 2000 }
        riskItems: { type: array, items: { type: string, maxLength: 200 } }
        recommendations: { type: array, items: { type: string, maxLength: 200 } }
        assessorId: { type: string, maxLength: 64 }
        assessedAt: { type: string, format: date-time }
    AssessmentReportReviewRequest:
      type: object
      required: [reviewStatus]
      properties:
        reviewStatus: { type: string, enum: [approved, rejected] }
    ApiAssessmentReport:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/AssessmentReport' }
...
    AssessmentReportPage:
      type: object
      properties:
        list: { type: array, items: { $ref: '#/components/schemas/AssessmentReport' } }
        pageNo: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, format: int64 }
    ApiAssessmentReportPage:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/AssessmentReportPage' }
...
    Device:
      type: object
      properties:
        id: { type: string }
        deviceName: { type: string }
        deviceType: { type: string }
        deviceSn: { type: string }
        elderlyId: { type: string, nullable: true }
        bindingStatus: { type: string, enum: [bound, unbound] }
        onlineStatus: { type: string, enum: [online, offline] }
        lastReportAt: { type: string, format: date-time, nullable: true }
        firmwareVersion: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DeviceCreateRequest:
      type: object
      required: [deviceName, deviceType, deviceSn]
      properties:
        deviceName: { type: string, maxLength: 100 }
        deviceType: { type: string, maxLength: 50 }
        deviceSn: { type: string, maxLength: 100 }
        firmwareVersion: { type: string, maxLength: 20 }
    DeviceUpdateRequest:
      type: object
      properties:
        deviceName: { type: string, maxLength: 100 }
        deviceType: { type: string, maxLength: 50 }
        firmwareVersion: { type: string, maxLength: 20 }
        onlineStatus: { type: string, enum: [online, offline] }
    DeviceBindRequest:
      type: object
      required: [deviceId, elderlyId]
      properties:
        deviceId: { type: string, maxLength: 64 }
        elderlyId: { type: string, maxLength: 64 }
    DeviceDataReportRequest:
      type: object
      required: [metrics]
      properties:
        reportedAt: { type: string, format: date-time }
        metrics:
          type: object
          minProperties: 1
          additionalProperties: true
    DeviceDataReport:
      type: object
      properties:
        id: { type: string }
        deviceId: { type: string }
        reportedAt: { type: string, format: date-time }
        metrics: { type: object, additionalProperties: true }
    ApiDevice:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Device' }
...
    DevicePage:
      type: object
      properties:
        list: { type: array, items: { $ref: '#/components/schemas/Device' } }
        pageNo: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, format: int64 }
    ApiDevicePage:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/DevicePage' }
    ApiDeviceReport:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/DeviceDataReport' }
    ApiDeviceReportList:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/DeviceDataReport' } }
...
    AnalysisRequest:
      type: object
      required: [scene, metrics]
      properties:
        scene: { type: string, enum: [care_decision_analysis] }
        tenantId: { type: string }
        regionCode: { type: string }
        timeRange:
          type: object
          properties:
            start: { type: string }
            end: { type: string }
        metrics:
          type: object
          required: [elderlyCount, warningCount, unhandledWarningCount, highRiskPopulationCount, deviceOnlineRate]
          properties:
            elderlyCount: { type: integer, format: int64, minimum: 0 }
            warningCount: { type: integer, format: int64, minimum: 0 }
            unhandledWarningCount: { type: integer, format: int64, minimum: 0 }
            highRiskPopulationCount: { type: integer, format: int64, minimum: 0 }
            deviceOnlineRate: { type: number, format: double, minimum: 0, maximum: 1 }
        dimensions:
          type: object
          properties:
            groupBy: { type: array, items: { type: string } }
        constraints:
          type: object
          properties:
            language: { type: string }
            outputFormat: { type: string }
            maxInsightCount: { type: integer, minimum: 1, maximum: 10 }
            mustInclude: { type: array, items: { type: string } }
...
    ApiAnalysis:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/AnalysisResult' }
...